requirements.txt
fastapi
uvicorn
numpy
tensorflow
torch
librosa
soundfile
python-dotenv

